+++ runner/android_junit_runner/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 **New Features**
 
+* Evaluate annotation with @CustomFilter on test class
+
 **Breaking Changes**
 
 **API Changes**